backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
pydantic==2.10.3
python-multipart==0.0.18
python-dotenv==1.0.1
google-genai==1.43.0
pypdf==5.1.0
sse-starlette==2.1.3
